safari-600.7-branch/LayoutTests/sputnik/Conformance/07_Lexical_Conventions/7.5_Tokens/7.5.2_Keywords/S7.5.2_A1.20-expected.txt

CONSOLE MESSAGE: line 76: SyntaxError: Unexpected token '='. Expected a block statement as body of a try statement.
S7.5.2_A1.20

PASS Expected parsing failure

TEST COMPLETE
